Moved FileStore under the adapters directory
ref https://linear.app/ghost/issue/HKG-1701

- Adapter-manager resolves via `require(<paths>/<type>/<class>)`, so FileStore has to live at `adapters/redirects/` for the registration to find it.
- Now `extends RedirectsStoreBase` with `super()` to pass the `instanceof` gate. The .d.ts shim alongside the JS base class lets the TS adapter default-import it under `esModuleInterop` (TS won't resolve a `.js` module without `allowJs`).
- Test relocated to `test/unit/server/adapters/redirects/` to match the layout `S3Storage` already follows. Contract helper stays put — moving it would drag the in-memory-store test along with no real benefit.
